Extract utility-like functions to separate file from Globe component
